Restructure to use Hcaptcha trait, HcaptchaRequest, HcaptchaResponse and HcaptchaErrror
