+++ Cargo.toml
@@ -24,6 +24,7 @@ enterprise = []
 trace = ["tracing"]
 
 [dependencies]
+async-trait = "0.1.50"
 hex = { version = "0.4", optional = true }
 reqwest = { version = "0.11.3", features = ["json"] }
 serde_derive = "1.0.117"
